Refactor the structur to represent the layers
