+++ src/presentation/pages/test/products-page/ProductsPage.helpers.ts
@@ -1,6 +1,6 @@
 import { waitFor, within, screen } from "@testing-library/react";
 import { expect } from "vitest";
-import { RemoteProduct } from "../../../api/StoreApi";
+import { RemoteProduct } from "../../../data/api/StoreApi";
 import { userEvent } from "@testing-library/user-event";
 
 export const verifyHeader = (headerRow: HTMLElement) => {
